+++ components/core/include/memfault/core/compiler_iar.h
@@ -32,6 +32,7 @@ extern "C" {
 #define MEMFAULT_GET_LR(_a) __asm volatile ("mov %0, lr" : "=r" (_a))
 #define MEMFAULT_GET_PC(_a) __asm volatile ("mov %0, pc" : "=r" (_a))
 #define MEMFAULT_PUT_IN_SECTION(x) _Pragma(MEMFAULT_QUOTE(location=x))
+#define MEMFAULT_BREAKPOINT(val) __asm volatile ("bkpt %0" : : "i"(val))
 
 #define MEMFAULT_STATIC_ASSERT(cond, msg) static_assert(cond, msg)