+++ components/demo/src/memfault_demo_cli_aux.c
@@ -15,4 +15,4 @@
 MEMFAULT_ALIGNED(4) static uint8_t s_test_buffer[16];
 void *g_memfault_unaligned_buffer = &s_test_buffer[1];
 // Also jump through some more hoops to trick the compiler into executing a bad function
-void (*g_bad_func_call)(void) = (void *)0xbadcafe;
+void (*g_bad_func_call)(void) = (void (*)(void))0xbadcafe;