+++ components/demo/src/memfault_demo_cli_trace_event.c
@@ -13,7 +13,10 @@
 int memfault_demo_cli_cmd_trace_event_capture(int argc, char *argv[]) {
   // For more information on user-defined error reasons, see
   // the MEMFAULT_TRACE_REASON_DEFINE macro in trace_reason_user.h .
-
+#if defined(MEMFAULT_TRACE_REASON_USER_DEFS_FILE)
+  MEMFAULT_TRACE_EVENT(Unknown);
+#else
   MEMFAULT_TRACE_EVENT(MemfaultDemoCli_Error);
+#endif
   return 0;
 }