+++ ports/zephyr/common/memfault_platform_core.c
@@ -11,18 +11,12 @@
 
 #include "memfault/core/compiler.h"
 #include "memfault/core/event_storage.h"
-#include "memfault/panics/reboot_tracking.h"
 #include "memfault/core/trace_event.h"
+#include "memfault/panics/reboot_tracking.h"
 
 #if !defined(CONFIG_MEMFAULT_EVENT_STORAGE_SIZE)
 #define CONFIG_MEMFAULT_EVENT_STORAGE_SIZE 100
 #endif
-
-void memfault_platform_halt_if_debugging(void) {
-  if (CoreDebug->DHCSR & CoreDebug_DHCSR_C_DEBUGEN_Msk) {
-    __asm("bkpt");
-  }
-}
 
 void memfault_platform_reboot(void) {
   const int reboot_type_unused = 0; // ignored for ARM